requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
chromadb==0.5.5
sentence-transformers==3.1.1
httpx==0.27.2
